Произведен небольшой рефакторинг. Удалены окна MainWindow переделано под  TabControl
