--- DitsApp/View/EquipmentWindow.xaml
@@ -1,26 +1,0 @@
-﻿<Window x:Class="DitsApp.View.EquipmentWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:DitsApp.View"
-        mc:Ignorable="d"
-        Title="EquipmentWindow" Height="450" Width="800">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <DataGrid x:Name="DataGridEquipment"
-                  Grid.Row="0">
-        </DataGrid>
-        <Button x:Name="ButtonAddEquipment"
-                Grid.Row="1"
-                    Content="Add New Equipment"
-                    Height="65"
-                    Width="180"
-                   VerticalAlignment="Top"
-                    Margin="10" Click="ButtonAddEquipment_Click"
-                    />
-    </Grid>
-</Window>
