--- DitsApp/View/EventsWindow.xaml
@@ -1,15 +1,0 @@
-﻿<Window x:Class="DitsApp.View.EventsWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:DitsApp.View"
-        mc:Ignorable="d"
-        Title="EventsWindow" Height="450" Width="800">
-    <Grid>
-        
-        <DataGrid x:Name="DataGridEvents">
-            
-        </DataGrid>
-    </Grid>
-</Window>
